+++ http/middleware.go
@@ -1,12 +1,13 @@
 package http
 
 import (
+	"log/slog"
 	"net/http"
+	"runtime/debug"
 	"time"
 
 	"go.strv.io/net"
 	"go.strv.io/net/internal"
-	"go.strv.io/net/logger"
 )
 
 const (
@@ -40,10 +41,26 @@ func RequestIDMiddleware(f RequestIDFunc) func(http.Handler) http.Handler {
 	}
 }
 
+type RecoverMiddlewareOptions struct {
+	enableStackTrace bool
+}
+
+type RecoverMiddlewareOption func(*RecoverMiddlewareOptions)
+
+func WithStackTrace() RecoverMiddlewareOption {
+	return func(opts *RecoverMiddlewareOptions) {
+		opts.enableStackTrace = true
+	}
+}
+
 // RecoverMiddleware calls next handler and recovers from a panic.
 // If a panic occurs, log this event, set http.StatusInternalServerError as a status code
 // and save a panic object into the response writer.
-func RecoverMiddleware(l logger.ServerLogger) func(http.Handler) http.Handler {
+func RecoverMiddleware(l *slog.Logger, opts ...RecoverMiddlewareOption) func(http.Handler) http.Handler {
+	options := RecoverMiddlewareOptions{}
+	for _, o := range opts {
+		o(&options)
+	}
 	return func(next http.Handler) http.Handler {
 		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 			defer func() {
@@ -56,10 +73,14 @@ func RecoverMiddleware(l logger.ServerLogger) func(http.Handler) http.Handler {
 					rw.SetPanicObject(re)
 					rw.WriteHeader(http.StatusInternalServerError)
 
-					l.With(
-						logger.Any("err", re),
-						logger.Any(requestIDLogFieldName, net.RequestIDFromCtx(r.Context())),
-					).Error("panic recover", nil)
+					logAttributes := []slog.Attr{
+						slog.String(requestIDLogFieldName, net.RequestIDFromCtx(r.Context())),
+						slog.Any("error", re),
+					}
+					if options.enableStackTrace {
+						logAttributes = append(logAttributes, slog.String("stack_trace", string(debug.Stack())))
+					}
+					l.LogAttrs(r.Context(), slog.LevelError, "panic recover", logAttributes...)
 				}
 			}()
 			next.ServeHTTP(w, r)
@@ -77,7 +98,7 @@ func RecoverMiddleware(l logger.ServerLogger) func(http.Handler) http.Handler {
 //   - Panic object if exists
 //
 // If the status code >= http.StatusInternalServerError, logs with error level, info otherwise.
-func LoggingMiddleware(l logger.ServerLogger) func(http.Handler) http.Handler {
+func LoggingMiddleware(l *slog.Logger) func(http.Handler) http.Handler {
 	return func(next http.Handler) http.Handler {
 		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 			rw, ok := w.(*internal.ResponseWriter)
@@ -90,26 +111,24 @@ func LoggingMiddleware(l logger.ServerLogger) func(http.Handler) http.Handler {
 			statusCode := rw.StatusCode()
 			requestID := net.RequestIDFromCtx(r.Context())
 
-			ld := LogData{
+			ld := RequestData{
 				Path:               r.URL.EscapedPath(),
 				Method:             r.Method,
 				RequestID:          requestID,
 				Duration:           time.Since(requestStart),
 				ResponseStatusCode: statusCode,
-				Err:                rw.ErrorObject(),
-				Panic:              rw.PanicObject(),
 			}
 
 			if statusCode >= http.StatusInternalServerError {
-				WithData(l, ld).Error("request processed", nil)
+				withRequestData(l, rw, ld).Error("request processed")
 			} else {
-				WithData(l, ld).Info("request processed")
+				withRequestData(l, rw, ld).Info("request processed")
 			}
 		})
 	}
 }
 
-// LogData contains processed request data for logging purposes.
+// RequestData contains processed request data for logging purposes.
 // Path is path from URL of the request.
 // Method is HTTP request method.
 // Duration is how long it took to process whole request.
@@ -117,30 +136,34 @@ func LoggingMiddleware(l logger.ServerLogger) func(http.Handler) http.Handler {
 // RequestID is unique identifier of request.
 // Err is error object containing error message.
 // Panic is panic object containing error message.
-type LogData struct {
+type RequestData struct {
 	Path               string
 	Method             string
 	Duration           time.Duration
 	ResponseStatusCode int
 	RequestID          string
-	Err                error
-	Panic              any
 }
 
-// WithData returns logger with filled fields.
-func WithData(l logger.ServerLogger, ld LogData) logger.ServerLogger {
-	l = l.With(
-		logger.Any("method", ld.Method),
-		logger.Any("path", ld.Path),
-		logger.Any("status_code", ld.ResponseStatusCode),
-		logger.Any("request_id", ld.RequestID),
-		logger.Any("duration_ms", ld.Duration.Milliseconds()),
-	)
-	if ld.Err != nil {
-		l = l.With(logger.Any("err", ld.Err.Error()))
+func (r RequestData) LogValue() slog.Value {
+	attr := []slog.Attr{
+		slog.String("id", r.RequestID),
+		slog.String("method", r.Method),
+		slog.String("path", r.Path),
+		slog.Int("status_code", r.ResponseStatusCode),
+		slog.Duration("duration_ms", r.Duration),
 	}
-	if ld.Panic != nil {
-		l = l.With(logger.Any("panic", ld.Panic))
+	return slog.GroupValue(attr...)
+}
+
+// withRequestData returns slog with filled fields.
+func withRequestData(l *slog.Logger, rw *internal.ResponseWriter, rd RequestData) *slog.Logger {
+	errorObject := rw.ErrorObject()
+	panicObject := rw.PanicObject()
+	if errorObject != nil {
+		l = l.With("error", errorObject)
 	}
-	return l
+	if panicObject != nil {
+		l = l.With("panic", panicObject)
+	}
+	return l.With("request", rd)
 }